+++ packages/template/src/lib/stack-app.ts
@@ -130,7 +130,8 @@ function getDefaultSuperSecretAdminKey() {
 }
 
 function getDefaultBaseUrl() {
-  return getPublicEnvVar("NEXT_PUBLIC_STACK_API_URL") || getPublicEnvVar("NEXT_PUBLIC_STACK_URL") || defaultBaseUrl;
+  const url = getPublicEnvVar("NEXT_PUBLIC_STACK_API_URL") || getPublicEnvVar("NEXT_PUBLIC_STACK_URL") || defaultBaseUrl;
+  return url.endsWith('/') ? url.slice(0, -1) : url;
 }
 
 export type StackClientAppConstructorOptions<HasTokenStore extends boolean, ProjectId extends string> = {
@@ -1571,7 +1572,6 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
     if (typeof window === "undefined") {
       throw new Error("callOAuthCallback can currently only be called in a browser environment");
     }
-
     this._ensurePersistentTokenStore();
     let result;
     try {
@@ -1581,8 +1581,10 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
     } catch (e) {
       if (e instanceof KnownErrors.InvalidTotpCode) {
         alert("Invalid TOTP code. Please try signing in again.");
+        return false;
+      } else {
+        throw e;
       }
-      throw e;
     }
     if (result.status === 'ok' && result.data) {
       await this._signInToAccountWithTokens(result.data);