+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/example-demo-app
 
+## 2.7.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.14
+  - @stackframe/stack-ui@2.7.14
+  - @stackframe/stack@2.7.14
+
 ## 2.7.13
 
 ### Patch Changes